Tasks C++ Wrapper
Summary: Implements rtos.h as a C++ wrapper. Task specific actions are a part of a Task class, the rest of the functions were simply added to the pros namespace.

Test Plan:
 [x] Create a task and interact with it as a C++ object
Test code:
```
void _task_fn(void* ign) {
	while (true) {
		std::cout << "Hello from task" << std::endl;
		pros::Task::delay(20);
	}
}

void initialize() {
	lcd_initialize();
	lcd_set_text(1, "Hello Beta Testers!");

	lcd_register_btn1_cb(on_left_button);
	pros::Task* task = new pros::Task(_task_fn);
	task_delay(600);
 	delete task;
	puts("here");
}
```
Results in:
```
hello from task
hello from task
hello from task
hello from task
hello from task
hello from task
hello from task
hello from task
hello from task
hello from task
hello from task
hello from task
hello from task
hello from task
hello from task
hello from task
hello from task
hello from task
hello from task
hello from task
hello from task
hello from task
hello from task
hello from task
hello from task
hello from task
hello from task
hello from task
hello from task
hello from task
here
```
 [x] Compiles

Reviewers: O1 The Ori, berman5, brookea

Reviewed By: O1 The Ori, berman5, brookea

Subscribers: brookea, berman5

Tags: #zorp

Differential Revision: http://phabricator.purduesigbots.com/D65

+++ include/rtos/task.h
@@ -307,8 +307,8 @@ is used in assert() statements. */
  */
 #if( configSUPPORT_DYNAMIC_ALLOCATION == 1 )
 	task_t task_create(task_fn_t task_code, void* const param,
-	                   uint32_t priority, const uint16_t stack_size,
-	                   const char* const name); 
+	                   uint32_t priority, const size_t stack_size,
+	                   const char* const name);
 #endif
 
 /**
